整合

+++ app/src/main/java/com/nexuslink/presenter/alterpresenter/AlterPresenter.java
@@ -15,7 +15,7 @@ import com.nexuslink.ui.view.BaseView;
 public abstract class AlterPresenter extends BasePresenter<AlterModel,AlterView> {
     public abstract void getUserInfo(int uId);
 
-    public abstract void changeUserInfo(int uId,char uGender, float uHeight, float uWeight);
+    public abstract void changeUserInfo(int uId,char uGender, int uHeight, int uWeight);
 
     public abstract void changeNickName(int uId,String uName);
